Ship CDN browser bundle with thin OpenUIDevtools host wrapper.
Move the full widget UI into dist/devtools.browser.js and keep the npm package as a thin CDN loader that forwards host React and all props.
