+++ docs/content/docs/api-reference/devtools.mdx
@@ -11,7 +11,19 @@ Development-only UI widget for OpenUI apps. Renders a floating button that opens
   className="w-full rounded-lg border"
 />
 
-## Install
+## Installation
+
+### CDN (via `react-lang`)
+
+If your app uses [`@openuidev/react-lang`](/docs/api-reference/react-lang), the widget shows up automatically in development. `react-lang` calls this package's thin helper, which fetches the browser build from jsDelivr (pinned to major `0`).
+
+#### CSP
+
+`script-src` must allow `cdn.jsdelivr.net` for the fetch to succeed. If it's blocked, the widget silently fails to appear – the rest of the app is unaffected.
+
+### Package Installation
+
+Install the package and render `<OpenUIDevtools />` when you want custom props. The component is still a thin CDN wrapper — your props (`theme`, `position`, …) are forwarded into the fetched widget. A manually mounted instance always wins over the auto-mount:
 
 ```bash tab="pnpm" tab-group="pkg"
 pnpm add -D @openuidev/devtools
@@ -40,7 +52,9 @@ function App() {
   return (
     <>
       {/* your app */}
-      {process.env.NODE_ENV === "development" && <OpenUIDevtools />}
+      {process.env.NODE_ENV === "development" && (
+        <OpenUIDevtools theme="dark" position="bottom-left" maxEvents={100} />
+      )}
     </>
   );
 }
@@ -53,7 +67,7 @@ function App() {
 | `enabled`         | dev-only         | Force the widget on/off.                                         |
 | `position`        | `"bottom-right"` | Corner for the toggle button: `top-left`/`top-right`/`bottom-*`. |
 | `maxEvents`       | `50`             | How many events to keep; oldest are dropped first.               |
-| `errorsOnly`      | `true`           | Capture only error/warning events, or every event.               |
+| `errorsOnly`      | `false`          | Initial state of the "errors only" display filter.               |
 | `autoOpenOnError` | `true`           | Initial state of the drawer's "auto-open on error" checkbox.     |
-| `bus`             | shared singleton | An `Observability` instance to listen to.                        |
 | `theme`           | `"light"`        | `"light"` or `"dark"` theme for the drawer.                      |
+| `cdnMajor`        | `@latest`        | Pin the CDN package major (`0` → `@0`). Omit for `@latest`.      |
