+++ docs/content/docs/api-reference/index.mdx
@@ -25,7 +25,7 @@ The OpenUI SDK is split into packages that build on each other:
 
 - **`@openuidev/browser-bundle`** — Prebuilt browser bundle for CDN, iframe, and no-build integrations. It packages the renderer, UI library, React, ReactDOM, and styles into script-tag-friendly assets.
 
-- **`@openuidev/devtools`** — Development-only floating widget that surfaces the events captured by `@openuidev/observability`, with error messages and stack traces.
+- **`@openuidev/devtools`** — Development-only floating widget that surfaces the events captured by `@openuidev/observability`. Auto-mounts from a CDN in `react-lang` apps.
 
 - **`@openuidev/cli`** — Command-line tool for scaffolding new OpenUI chat apps and generating system prompts or JSON schemas from library definitions.
 
@@ -97,7 +97,7 @@ The OpenUI SDK is split into packages that build on each other:
     CDN, iframe, and no-build browser bundle for the renderer, UI library, React, and styles.
   </Card>
   <Card title="@openuidev/devtools" href="/docs/api-reference/devtools">
-    Development-only floating widget surfacing captured events with error messages and stack traces.
+    Development-only floating widget. Auto-mounts from a CDN in development for react-lang apps.
   </Card>
   <Card title="@openuidev/cli" href="/docs/api-reference/cli">
     openui create (scaffold a Next.js app) and openui generate (system prompt + library spec from a
