+++ packages/devtools/package.json
@@ -26,7 +26,7 @@
   },
   "scripts": {
     "test": "vitest run --passWithNoTests",
-    "build": "tsdown",
+    "build": "tsdown && node scripts/build-browser.mjs",
     "watch": "tsdown --watch",
     "typecheck": "tsc --noEmit",
     "lint:check": "eslint ./src",
@@ -74,6 +74,7 @@
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
+    "esbuild": "^0.25.12",
     "jsdom": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
