+++ packages/devtools/src/browser-shims/react.ts
@@ -1,0 +1,17 @@
+import { requireReact } from "./slots";
+
+const react = requireReact();
+
+export const {
+  useState,
+  useEffect,
+  useCallback,
+  useMemo,
+  useRef,
+  useContext,
+  createContext,
+  createElement,
+  forwardRef,
+  Component,
+  Fragment,
+} = react;
