+++ packages/devtools/src/index.ts
@@ -1,4 +1,4 @@
 "use client";
 
 export { OpenUIDevtools, type OpenUIDevtoolsProps } from "./OpenUIDevtools";
-export type { ColorMode } from "./theme";
+export type { ColorMode, DevtoolsPosition } from "./types";
